Use Knapsack.logger for runner output
Currently the different runners use `puts` to directly print their output to
STDOUT.

This commit changes them to use Knapsack.logger at INFO level. By default this
will still print to STDOUT as before, but allows users to send this output
elsewhere, or suppress it with a higher log level if desired.