Log the time offset warning at INFO if time not exceeded
Currently, at the end of every suite run, the node's time offset warning is
logged at WARN level. However, in the event that the allowed time offset is
not exceeded, this does not require user action.

This change modifies the reporting calls to log the message at INFO level if
the time offset is not exceeded, and WARN if it is exceeded. This allows a
user to set their log level to WARN, and receive the message only when they
need to rebalance their knapsack timing report.